add getReadersTakenBooks & getReadersReservedBooks, fix assign and reserve Book

+++ src/main/java/ua/karazin/interfaces/ProjectLibrary/dto/ReadersBookCopyDTO.java
@@ -3,7 +3,7 @@ package ua.karazin.interfaces.ProjectLibrary.dto;
 import java.util.Date;
 
 public record ReadersBookCopyDTO(
-        Integer isbn,
+        Long isbn,
         Integer copyId,
         String title,
         Object authors,
